T-1-extra: Add Main class with main function to make TicketService standalone service
