+++ src/TicketService.java
@@ -5,18 +5,10 @@ import java.util.List;
 
 public class TicketService {
 
-    private List<Ticket> tickets = new ArrayList<>(10);
+    private final List<Ticket> tickets;
 
-    public static void main(String[] args) {
-        Ticket ticket1 = new Ticket();
-        Ticket ticket2 = new Ticket(196, "Pantheon",
-                101, LocalDateTime.parse("2024-06-30T19:30:00").atZone(ZoneId.of("Asia/Almaty")).
-                toInstant().toEpochMilli(), false, "A", 10.5);
-        Ticket ticket3 = new Ticket("Pantheon", 101, LocalDateTime.parse("2024-06-30T19:30:00").
-                atZone(ZoneId.of("Asia/Almaty")).
-                toInstant().toEpochMilli());
-
-
+    public TicketService(int capacity) {
+        this.tickets = new ArrayList<>(capacity);
     }
 
     public void addTicket(Ticket ticket) {
